Separing control flow under trace VS replacement with static Python control flow
