+++ src/qrisp/jasp/interpreter_tools/interpreters/cl_func_interpreter.py
@@ -73,7 +73,7 @@ from qrisp.jasp import (
     insert_outvalues,
 )
 from qrisp.jasp.interpreter_tools.abstract_interpreter import ContextDict
-from qrisp.jasp.interpreter_tools.interpreters.control_flow_interpretation import (
+from qrisp.jasp.interpreter_tools.interpreters.traced_control_flow_interpretation import (
     evaluate_scan_under_trace,
     flatten_signature,
     unflatten_signature,
